Fix #6208, #6605: Properly support toplevel defs in the IDE
Toplevel defs are enclosed in an object named `<source>$package` by
`Desugar#packageDef`, this didn't work correctly in the IDE because the
name of the VirtualFile was the whole URI of the file. In particular
this lead to double-vision problems ("cannot merge ...") when the
toplevel definition was also present on the classpath.

Also one of the overload of Driver#compile to Driver#compileFromString
for clarity and give better names to VirtualFile it creates since they
could also leak into the names of generated objects.