Ignore all kinds of ProtoTypes instead of just AppliedProtos
In blocks, conditionals and matches we mapped applied prototypes to
WildcardTypes in order to force eta expansions. With extension methods
the same problem arises for SelectionProtos as well. Furthermore,
it makes to sense to propagate a ViewProto into a local context. So the
consistent thing to do is to map all kinds of ProtoTypes to WildcardTypes
in these situations.

This fails one line in run/implicitsFuns.scala, which, however is of dubious value.
(The line requires a manually inserted `apply` to propagate a `given` argument
into an if-then-else, circumvening the previous "no applies propagated" strategy.