Rely on variance to typecheck more uses of wildcards
If we can't capture the wildcard argument, we might still be able to
return true to the subtype check by approximating the wildcard based on
variance.

This leads to a regression in t5729 were two overloads are now
considered ambiguous, but the only reason they were not ambiguous before
was that the subtype test failed when it shouldn't, so this seems
acceptable.