Improve handling of nameSpan
Even with the fix to enum spans, we could still get a case where
a point to a name is missing (maybe because there is a long comment
between the start of the definition and the name). The current fix
makes the computation succeed more often, but it is not perfect.