Fix #6603: Make privateWithin force less
At this point, ModuleCompleters and SymbolLoaders are the only two kinds
of completers which need to set privateWithin, this is now enforced by
the setter (which is therefore renamed to `setPrivateWithin` to make it
clearer that it doesn't just set a field) and taken advantage of by the
getter to avoid forcing unless necessary.

This is the last piece needed to finally fix #6603. Unfortunately, the
cumulative changes to completions done so far have caused cyclic
reference errors to pop up in new situations, including bootstrapping
Dotty itself. This is fixed by the remaining commits in this PR.