Partial fix for inlining with opaque types
When compiling shapeless, we saw an error message
```
[error] -- [E007] Type Mismatch Error: /home/miles/projects/shapeless-ng/core/src/test/scala/shapeless/type-classes.scala:53:80
[error] 53 |  inline def derived[A] given (gen: K0.ProductGeneric[A]): Monoid[A] = monoidGen
[error]    |                                                                                ^
[error]    |Found:    shapeless.K0.ProductInstances[shapeless.Monoid, shapeless.ISB]
[error]    |Required: shapeless.K0.ProductInstances[shapeless.Monoid, shapeless.ISB]
```
With -Yprint-debug that became clearer:
```
|Found:    K0.this.ProductInstances[[A] => shapeless.this.Monoid[A], shapeless.this.ISB]
|Required: shapeless.this.K0.ProductInstances[[A] => shapeless.this.Monoid[A],
```
This commit ensures that the type K0.this.ProductInstances does not show up outside of `K0`.
There were two places in the inliner where `M.this` for static `M` was kept as is. Normally
that's OK, but it does matter if `M` contains opaque types.

This is a partial fix only. Compared to the original shapeless code, test pos/shapeless.scala also
contains two casts (on lines 65 and 68), which would lead to inlining failures if missing.

Unfortunately, this second problem looks impossible to fix in general. We cannot simply move
code compiled inside an opaque alias scope to outside this scope. The compiler can try to insert
some casts to make it work, but I see no general proven way to do so, because opaque types are
explained through type abstraction instead of through implicit conversions.