Add effpi to our community-build
See https://github.com/alcestes/effpi for more information. This project
is interesting to have in the community build since it exercises various
Dotty features (implicit/dependent function types, unions, type lambdas,
...) and even contains the first real Dotty compiler plugin! We compile
this plugin but do not run it as part of the community build since it
depends on native binaries (it calls into the mcrl2 model checker).