Fix #6142: remove unreasonable type change for _
The logic was introduced in f618e47, which is hard to justify.
The problem is exhibited in tests/pos/i6142.scala, where
we have code lik the following:

    implicit val _: scala.quoted.Type = ...
    {
      def $anon$ = _
      closure($anon$)
    }

Changing the type of `_` will break LambdLift (`_` is not free anymore),
and the compiler crashes at backend when generating code for the unfound local variable `_`,
whose symbol is `scala.quoted.Type`.