Non-parametric derived instances now lazy vals
Previously all derived instances would be compiled as given defs, which
have caching logic added, where possible, in CacheAliasImplicits. These
caches are unsynchronized, which is a reasonable default in
circumstances where the progammer has the option to manually insert a
lazy val or some other form of synchronization. That option isn't
available for derived instances, however, because these givens aren't
written explicitly. In this case the more appropriate default is to
compile non-parametric derived instances as given lazy vals.