Update docs
The new implementation does not optimize given aliases with pure paths as right hand sides to be
vals anymore. The reason is that even a pure path might be expensive to compute so we might want
the caching. And caching is cheap, so there's little downside. Therefore, it's attractive to go
with the simpler two way choice: leave RHS as is, or wrap it in a memo.