Drop per-run infrastructure
The per-run infrastructure caused a 10-20% performance hit. We could go back to the old
caching scheme but this is verbose and ugly. This commit drops the whole caching infra
stricture instead. The downside is that we need a new compiler if any symbol cached
by Definitions has been edited and should be reloaded.