Cross-compile the compiler with the 2.13 stdlib
We still need to support compiling it against the 2.12 stdlib until we
release a new reference compiler because the current reference compiler
cannot use the 2.13 stdlib. This is mostly not a problem, except that
this forces us to keep two versions of WeakHashSet, the same source file
cannot be used for both the 2.12 and 2.13 stdlib since `-=` and `+=` are
final in 2.13.