Fix classloader confusion in the REPL
The REPL constructed a classloader containing the user classpath but
having the classloader used to run the REPL as a parent, meaning that
the scala-library we actually used came from the JVM classpath, so of
course everything exploded when using a repl compiled with the 2.12
stdlib and running with the 2.13 stdlib. Fixed by having
Rendering#classpath be constructed without any parent, just based on the
user classpath. This required some additional changes:

- We called ScalaRuntime#replStringOf using the ScalaRunTime class from
the scala-library used by the REPL, switch to calling it reflectively
through
- ReplDriver#withRedirectedOutput used `Console.withOut/Err` but that only
affects the Console object in the current classloader, not the one in
the user classloader, use `System.setOut/Err` instead (the class
`System` comes from the system classloader so there's only one of it).
- The macro tests now need to be run with the compiler on the user
classpath, this required some refactoring (This would be simpler to
handle if ReplTest didn't extend ReplDriver, but I'll let someone else
do that refactoring)