Take number of type parameters into account when choosing an overload
In Tasty, pickling a selection will pickle its signature using the
SIGNED tag to be able to distinguish overloads, this usually works but
it turns out that some overloads have the same signature (see
tests/pos/poly-signature.scala, but this also affects `Set#++` in the
2.13 collections), resulting in ambiguous overload errors when
unpickled.

This commit solves this problem by also storing the number of type
parameters in the signature. The obvious way to implement this would be
to add a field of type Int to the `Signature` case class, but that
wouldn't scale to a future where methods may have multiple type
parameter lists, so instead we change the existing `paramsSig:
List[TypeName]` field to take a list of `TypeName | Int`.