Avoid memory leaks in Stream methods
`Stream.collect/collectFirst/find` kept references to the original
receiver of the call which prevented any part of the Stream from
getting garbage collected as soon as possible.

The previous implementation of `Stream.collect` claimed to avoid this
but this only worked insofar as the resulting Stream did not hold on
unnecessarily to any references to the original Stream. It still kept
a reference to the original receiver while iterating over non-matching
elements which can only be avoided with a tail-recursive
implementation.

Fixes https://github.com/scala/bug/issues/11443