Fix #6989: Make sure scalacLinkedClass finds companion when unpickling
scalacLinkedClass uses unforcedDecls which might miss a companion object when
unpickling. I tried to make scalacLinekdClass stricter, but ran into
problems. It turns out that it is finely balanced as it is.

The fix is to make sure the owner is completed before calling scalacLinkedClass
from an unpickler. I trued to do the same for the Scala2Unpickler but that ran
again into problems. So this is quite the mine-field...