Implement generic number literals
### Example:

```
val x: BigInt = 111111100000022222222222
```

### Also allow generic literals in patterns

Wrap generic literals in patterns in blocks, which force
expression evaluation. This allows to match (say) a scrutinee
of BigInt type against large numeric literals. But it does
not work if the scrutinee is of type `Any` because then
the expeected type for the pattern is missing.

This would be addressed by a feature that's still missing: patterns
of the form `<literal> : <type>`. We have to change the syntax of patterns
in general for this one.