Make `then` optional at the end of line
`then` is treated like `;`: It is inferred at the end of a line, if

 - a new line would otherwise be inferred
 - the next line is indented

Disambiguation with Scala-2 syntax is as follows:

The question is whether to classify

   if (A) B
     C

as a condition A followed by a statement B (old-style), or
as a condition (A) B followed by a then part C, and an inferred
`then` in-between (new-style)

(new-style) is chosen if B cannot start a statement, or starts with a leading
infix operator. Otherwise put, (new-style) is chosen if in
```
(A)
B
```
no newline would be inserted. (old-style) is chosen otherwise.

This means that some otherwise legal conditions still need a `then` at the end of a line.
If (old-style) is eventually deprecated and removed, we can drop this restriction.