Fix #7249: Search for hidden givens only at whole type
Don't try to find a hidden implicit for a part of type type that
was searched for, unless that part is fully defined.

The reason is that any type variables in a subtype might be associated
with constraints that are lost at the point of issuing a the hidden
givens addendum.

The commit also contains a tweak to "isFullyDefined" which makes
an uninstantiated type variable that is not bound in the current
constraint count as "not fully defined", where previously it was
"fully defined".