Suspend callers of macros compiled in the same run
Allows macros to be defined and called in the same project. The idea is to suspend
the callers of a macro (and their upstream dependencies), compile the sourcefile(s)
containing the macro(s), and then compile the suspended files in a second run.

The scheme detects and flags as errors calls to macros in the same file and mutual
dependcies between different files which would not work with suspensions.

It needs to be complements by another mechanism that detects that class files loaded
by a macro class loader do not have assiciated source files that are currently compiled.