Fix-#4867 Keep Unions when explicit
Improves the inference of union types by preserving the
union when there is a type ascription with an union.

If the term has not been explicity ascribed with an union then
the existing semantics of joining the orType is maintained.

To determine whether an explicit union ascription exists, a lexical
check on the untyped tree is performed.