Disambiguate quoted and tasty Type
This ambiguity can be seen in the following
```scala
import scala.quoted._
def f[T: Type](given qctx: QuoteContext) = {
  import qctx.tasty._
  ...
  summon[Type[T]] // qctx.tasty.Type instead of scala.quoted.Type
  ...
}
```

To avoid the ambiguity we simply rename `scala.quoted.Type` to `TypeTag`.
From the 2 this is the one that conforms the least to the common expectation of
what is a `Type`. We chose `TypeTag` because it encapsulates the same kind
of information that `scala.reflect.TypeTag` did.