Revert invert type signature change
This aligns it with what 2.12 has. I was going to justify this with a
comment on how Dotty also can't infer the types, but as it turns out
Dotty can do so just fine, so now I need to justify this with a comment
on how much more difficult type-inference bugs are to fix.

Fixes scala/bug#11766.