Do not reset the GivenOrImplicit flags on implicits without explicit types
Also fix a neg test for i3067. After this modification to
the completion algorithm, implicit definitions without
explicit types stay visible in the implicit scope and hence,
the error of the missing implicit is not reporter. This
is fine, however, since the real error to be fixed is the
lack of an explicit type of the implicit value, not the
lack of the implicit value itself.