Fix i7407 test error message
FTTB it is a cyclic error. It requests the user to
fully specify the type of the method. Once they do so,
an intended "implicit not found" message appears.
Displaying the correct message directly requires adding
`Macro` flag to `FromStartFlags`, which is not correct.
Until we find a workaround, this solution is a reasonable
trade-off because it offers an unambiguous action (to
fully specify the type of the method) for the user to
take to see the correct error message.

The reason the cyclic error appears in first place is as
follows. Given `inline def g = qc`, we typecheck `qc`.
However, `qc` requires an implicit QuoteContext which
can only be synthetically generated for the bodies
of macros. Hence we need to see if `g` is a macro, and
to do so, we need to complete it. To complete it, we need
to typecheck its body, which yields the cyclic error.