Tighten flag checking rules on completion
Count a denotation as completing if it has the `Touched`
flag set. This assumption follows from the definition
of the flag `Touched` provided in its javadoc comment
in Flags.scala. Also, this commit unconditionally unsets
this flag for all denotations once they are completed. The
rationale is, again, to comply with the definition of
`Touched`.