Revert "Revert "Upgrade to ASM 7.0""
This reverts commit 5a747185826bf2b178389b6f48880eb2e5d252a7 which
itself reverted 0abd0769bf27682e1bd62e9532afdafce2c92dfb.

Since #5936 is not fixed yet, so we need to turn off one test in
semanticdb which uses parens in a class name. This is problematic, but
upgrading to ASM 7.0 still seems worth it since it allows running scalac
and dotty in the same classpath which should make it easier to test the
scalac tasty reader.