Introduce conditionally mutable flags wrt completion
Some flags do not fall under the definition of the
flags that are not mutated during completion. However,
in practice, they are almost never mutated on completion
and such an event is easy to predict. An example of such
a flag is `deferred` which is set only when compiling
opaque types – this is part of what makes opaque types work.
It is reasonable, hence, to make sure denotations are counted
as current even though they have some of such flags if
when querying these flags we can predict that they will not
change during completion.