Process Deferred flag last on unpickling
Now that Deferred is conditionally immutable,
if we query it along with other properly immutable flags,
the query may trigger completion. This will happen if the
symbol in question is also opaque and is a type.

The completion will be rightfully triggered since
the opaque mechanism does indeed mutate the
deferred flag on opaque types. However, if e.g.
we query `is(Method, butNot = Deferred)` on
`opaque type B`, we may be able to return `false`
on the query as soon as we see that the symbol in
question is not a method, without triggering the completion.

Too early completion is undesirable since completion of
opaque types modifies the self-type of its owner.
If the completion of an opaque type happens during unpickling
of the owner itself, the self-type of the owner may be
rewritten later in its completion.