Rename isCompleted and isCompleting methods of SymDenotation
isCompleted is now isCompletedOrStubbed and isCompleting –
isCompletingAndNotStubbed. This reflects the fact that sometimes
we use a hack where we set `info` on denotations to a non-lazy type
during completion to flip these predicates. This hack is possible
because both of these predicates depend on checking whether
the `info` is non-lazy.

However, if we make these predicates with these specific names
depend on the `info`'s type, their names became incorrect wrt
their semantics. `isCompleting` must be true iff the `Completing`
(formerly `Touched`) flag is set to true by definition of that flag.
`isCompleted` must be true if we both have a concrete `info` and
are not in process of being completed (since further modifications
may be done by the completion process, we cannot say a denotation is
completed unless this process has finished).

In principle we should not use the hack in question in first place
and have something more well-defined to avoid cyclic references.
The bare minimum we can do right now is to rename the predicates to
reflect what they do more precisely so that to avoid confusion when
reading the code.