Scala2x is an AfterLoadFlag
Scala2x is a flag set for all the Scala 2 compliled classes.
The information on what Scala version was used to compile the
class file is available from the bytes of the classfile itself.
Hence we cannot have this information from start, it will only
be available on load. See ClassfileParser.unpickleOrParseInnerClasses
for details on how the Scala version is determined.