On Scala 2 Unpickling, mark moduleRoot symbol as loading
`moduleRoot` is derived from moduleClassRoot which is
being loaded. Since it is derived during the loading
sequence of `moduleClassRoot`, we can argue that
`moduleRoot` is loaded during this sequence as well.
Hence we mark `moduleRoot` as loading by setting its
completer to a temporary value. We set the completer back
after the loading is done.