Refine SelectionProto#isMatchedBy
Previously, uninstantiated type variables as the prefix marked the prefix
as having unknown members, which meant that isMatchedBy returned true. Now
we do that only if the type variable has a bottom type as lower bound. This
reflects a planned change that a type variable with non-bottom lower bound
is instantiated to its lower bound when adapting against a selection proto.