Fix #6291: Rename Scala2 language setting to Scala2Compat
The Scala2 language option is misleading. On several occasions I met people who had the impression that it would be a mode where Scala 2 is fully supported but none of the Scala 3 features would be accepted. Now we use

 ```
 -language:Scala2Compat
 ```
as a command line option and

```scala
import language.Scala2Compat
```
as an import.