Disallow old anonymous extension syntax
Disallow the previous syntax
```
  given (x: T)
    ... extension methods
```
Require instead
```
  given extension (x: T)
    ... extension methods
```
This avoids the confusion with given instances for types in parentheses and makes
i7515.scala compile. To allow building with 0.20rc1, we allow the old extension syntax
if it comes with a name. E.g. the following is still allowed, but will be dropped
once we have bootstrapped with this commit present.
```
  given ops: (x: T)
    ... extension methods
```