Fix #7414: Better approximation for class references
What to do if there is an as-seen-from result of the form
```
  (Nothing .. T) # ref
```
i.a. a projection over a range? The fallback so far was the
range `(Nothing # ref .. T # ref)`. But if `ref` is a class,
we can probably replace it by widen(T) # ref instead. A type projection like that
is legal, and subsumes all possible types in the range.