Make fresh name generation per file instead of global
This makes the compiler output more stable and avoids overcompilation
during incremental compilation due to context bounds parameters changing
name. This is still not fine-grained enough as noted in the
documentation of CompilationUnit#freshNames.

To work properly, this requires us to always have a compilation unit set
in the context before compiling code, which is a good idea anyway (would
be nice to improve the API to make this harder to screw up).