Change -sourcepath to behave like -scansource -sourcepath
... and delete -scansource which is now unnecessary.

The difference between these two settings is that the former assumes
that a file A.scala contains a top-level class or object A, whereas the
latter will use the outline parser to find definitions in A.scala and
create symbols for them.

So far we used the former to compile the standard library and
dotty-library, and the latter for the IDE, but we might as well always
use the latter, it's one less code path and dotty-library now uses
top-level definitions (like `type IArray` in scala/IArray.scala) so the
assumptions of -sourcepath do not apply to it anymore.

(top-level definitions are still not handled correctly with -sourcepath
after this commit, this is fixed in a latter commit in this PR)

This change required moving WithBounds which was previously in the wrong
directory.