sourcepath: don't create a new run for outline parsing
The previous commit broke compilation of the standard library, because
we ended up calling `runContext` before the first context is set, when
it's still `null`. But we don't really need a new run just for creating
the symbol: the completer we set will call Run#lateCompile which will
take care of creating a new run anyway.