Report an error if toplevel denotations come from multiple files
Report an error if toplevel denotations the same name with come
from multiple files that have the same modification date.

The error was reported in some runs for toplevel/stale when run locally
since the files for A_1.scala and B_2.scala were generated too fast so they
ended up with the same modification date. B_2.scala has now been beefed
up to give the compiler more work to do, so that the race is avoided.