sourcepath: Handle end markers in outline parsing
Previously, outline parsing would have failed on indent1.scala with:

7 |  end inner
  |  ^
  |  misaligned end marker

because `OutlineParserCommon#skipBraces` does not update
`in.openEndMarkers`, but calls `in.nextToken` which eventually calls
`in.skipEndMarker` which checks that end markers are aligned.

This commit fixes this by adding a boolean in the scanner to turn off
end marker checks, this is used by the outline parser when skipping over
code.