Radix-Balanced Finger Tree Vectors
This is complete rewrite of the old scala.collection.immutable.Vector.
The design differs from the old one in one important aspect: Instead
of a single movable finger (the "display" pointers) it has a fixed
finger at each end. These fingers are part of the proper data
structure, not a transient optimization. There are no other ways to
get to the data in the fingers (it is not part of the main data
array), no focus point, and no stabilization. As a side-effect, the
data structure is completely immutable and no extra memory fences are
required. A second major difference lies in the implementation.
Whereas the old Vector uses a single class for collections of all
supported sizes, the new one is split into Vector0 to Vector6 for the
different dimensions of the main data array.