Put TupledFunction object in the `internal` package folder
Done to match the location of the source to its package.
The original source, thus, ends up split into two files in different
packages. Since a sealed trait is involved, a new constructor for
this trait is created to facilitate the access to this trait from
the source that was moved from the original file.