Don't use a classloader to read .tasty files from jars
The previous commit tried to fix some weird caching behaviors by closing
the classloader used to read a .tasty file, but it turns out the JVM is
dumb and that breaks any other in-use classloader that uses the same
jar. Give up on the whole thing and work directly with the underlying
ZipArchive to find the .tasty file (this just required adding a "parent"
field to ZipArchive#Entry).