Fix tastyBootstrap test to also compile tasty-core
tasty-core is a dotty project, so when bootstrapping it needs to be
recompiled and cannot be reused as is, the test did not fail so far
since we didn't break anything, but it could be made to break by bumping
the tasty version in TastyFormat.

Also use change the names of the output directories of lib, dotty1 and
dotty2 to match the group name.