Fix #7959: Don't check Java symbols for private leaks
Java does not have the same rules as Scala here. Also, there are problems
with package[private] which is translated to private[pkg]. In the test case,
the constructor is assumed to be visible up to the package (because it
is always annotated as such) which is further than the enclosing private class.