Docs: Reformulate match type reduction algorithm
I didn't understand that section on MT reduction, the rewritten version
seams to correspond more closely to what's implemented in the compiler.

I didn't get to review that text when it was first added to the docs,
but here are the parts that are giving me trouble:

- Which part of the code implements the "minimal instantiation" as
described in the text?

- What is `i` in the two "can-reduce" code blocks? Is it related to
T/T'?

- The "cannot possibly match" relation, which is central to the
algorithm, isn't defined anywhere.