Reject char literal of raw newline
Don't allow char literal to span a natural line
ending, but do allow `'\u000A'` to mean `'\n'`.
The spec was updated to reflect that meaning,
and the direction is to eliminate Unicode escapes
outside quotes.

Update the test rig so that `// anypos-error` can be
used to match any position, since the test cannot
be expressed on one line with a trailing comment.