Fix #8349: Generic signatures of generic arrays
When T[] is erased to Object, it should get a Java generic signature of
Object and not of T<>. Fixed by reusing the same logic we have in
TypeErasure which replaces a bunch of weird code imported from scalac.

The new behavior differs in one case from scalac: `Array[_ <: Byte]`
gets a generic signature of `byte[]` instead of `Object`, but this is
consistent with how we handle erasure (which I think is OK because as
far as I can tell, we can't construct a value of type `Array[Nothing]`
with Dotty).