TreeMap is serialization compat with 2.12.10 again
Explicitly lock down the SerialVersionUID and avoid having
scalac mangle the name of the `tree` field to make it non-private.

Reinstate serialization compatibility tests for this and other
collections that were temporarily disabled due to an interation
with -Xcheckinit. Now, we simply skip the test when we detect
that the build is using that option.