Don't check key equality in RedBlackTree
This was left over from 2.12 when they old key would be overwritten.
The behavior changed in https://github.com/scala/scala/pull/7481 to
comply with the new map/set rules in
https://github.com/scala/bug/issues/10415#issuecomment-442926971.

I would also call it wrong in 2.12 because equality should be defined
exclusively by the Ordering, but we should keep the current behavior
for compatibility reasons.