Run ignoredDeferred less often
RefChecks would invoke ignoreDeferred each time we hit an abstract member in a base class of a
checked class, which means almost all of the time. ignoreDeferred would in turn do time travel
to post-erasure. This causes inefficiencies because of compiler-context switches. It also exposed
the problem in i8425 that was fixed by the last commit.

The concrete failure for i8425 does go away with the change in this commit. However, the fix
in the last commit is still necessary since explicit outer/erasure races are a general problem
that could come back.