Optimize canDropAlias
Cache canDopAlias calls in TypeRefs.

Performance tests pointed to a ~1% showdown before, which is really in the noise.
However, stats showed that the number of iterations in the exists call of
`canDropAlias` is significant (a bit more than total member operations).
With this optimization, the number of calls gets reduced to 1% of what it was.